+++ lib/booqable/configurable.rb
@@ -48,6 +48,7 @@ module Booqable
                   :proxy,
                   :read_token,
                   :redirect_uri,
+                  :refresh_token_wrapper,
                   :single_use_token,
                   :single_use_token_algorithm,
                   :single_use_token_company_id,
@@ -81,6 +82,7 @@ module Booqable
           proxy
           read_token
           redirect_uri
+          refresh_token_wrapper
           single_use_token
           single_use_token_algorithm
           single_use_token_company_id